FIX: tri_tas, création du tableau générale

+++ src/triBulle/tri_bulle.h
@@ -1,6 +1,6 @@
 #ifndef _TRI_BULLE_H
 #define _TRI_BULLE_H
 
-void tri_bulle (int *tab, int tailletab);
+void tri_bulle (float *tab, int tailletab);
 
 #endif
