+++ src/triTas/tri_tas.h
@@ -1,6 +1,6 @@
 #ifndef _TRI_TAS_H
 #define _TRI_TAS_H
 
-void tri_tas (int *tab, int max);
+void tri_tas (float *tab, int max);
 
 #endif
